So it 'plays audio' with symphonia, but this blocks the main thread. I DO need to create a separate thread which processes audio commands.
